# openapi.yaml
openapi: 3.0.3
info:
  title: CRUD API
  description: A simple CRUD API for items
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /items:
    get:
      summary: Retrieve all items
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: objectId
                    name:
                      type: string
    post:
      summary: Create a new item
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "New Item"
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: objectId
                  name:
                    type: string
  /items/{id}:
    get:
      summary: Retrieve an item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: objectId
      responses:
        '200':
          description: An item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: objectId
                  name:
                    type: string
        '404':
          description: Item not found
    put:
      summary: Update an item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: objectId
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: objectId
                  name:
                    type: string
        '404':
          description: Item not found
    delete:
      summary: Delete an item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: objectId
      responses:
        '204':
          description: Item deleted
        '404':
          description: Item not found